+++ packages/micromark-extension-kbd-nested/CHANGELOG.md
@@ -2,14 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [0.3.1](https://github.com/shivjm/remark-extensions/compare/micromark-extension-kbd-nested-v0.3.0...micromark-extension-kbd-nested-v0.3.1) (2021-11-02)
+## 0.4.0 (2021-11-02)
+
+### ⚠ BREAKING CHANGES
+
+* **micromark-extension-kbd-nested:** update and improve README
+* **micromark-extension-kbd-nested:** add test suggested by README
+* **micromark-extension-kbd-nested:** add support for variable sequences
+
+### Features
+
+* **micromark-extension-kbd-nested:** add support for variable sequences ([d725d24](https://github.com/shivjm/remark-extensions/commit/d725d24190649dfe0e7cb8e656b4571482c3554d))
+
+
+### Tests
+
+* **micromark-extension-kbd-nested:** add test suggested by README ([cbac68b](https://github.com/shivjm/remark-extensions/commit/cbac68b9dd584715be36f793b7a660aa541b9c87))
+
+
+### docs
+
+* **micromark-extension-kbd-nested:** update and improve README ([362920b](https://github.com/shivjm/remark-extensions/commit/362920b09578598c945720c2bb9e24130fdf48fb))
+
+## 0.3.1 (2021-11-02)
 
 
 ### Bug Fixes
 
 * **micromark-extension-kbd-nested:** correctly handle escaped backslashes ([44c0515](https://github.com/shivjm/remark-extensions/commit/44c05153761e7d4af86b877f71e1f74952f3fb7b))
 
-## [0.3.0](https://github.com/shivjm/remark-extensions/compare/micromark-extension-kbd-nested-v0.2.1...micromark-extension-kbd-nested-v0.3.0) (2021-11-02)
+## 0.3.0 (2021-11-02)
 
 
 ### ⚠ BREAKING CHANGES
@@ -26,14 +48,14 @@ All notable changes to this project will be documented in this file. See [standa
 
 * **micromark-extension-kbd-nested:** preserve space inside `kbd` sequence ([9a6a0eb](https://github.com/shivjm/remark-extensions/commit/9a6a0eb6f654ba469891d20d192df327e86340be))
 
-### [0.2.1](https://github.com/shivjm/remark-extensions/compare/micromark-extension-kbd-nested-v0.2.0...micromark-extension-kbd-nested-v0.2.1) (2021-11-01)
+## 0.2.1 (2021-11-01)
 
 
 ### Bug Fixes
 
 * **micromark-extension-kbd-nested:** put micromark-util-symbol in `dependencies` ([749f300](https://github.com/shivjm/remark-extensions/commit/749f300964b95a286600b206fabdc8598c3d24e4))
 
-## [0.2.0](https://github.com/shivjm/remark-extensions/compare/micromark-extension-kbd-nested-v0.1.0...micromark-extension-kbd-nested-v0.2.0) (2021-11-01)
+## 0.2.0 (2021-11-01)
 
 
 ### ⚠ BREAKING CHANGES